requirements.txt
pandas==2.2.3
psycopg2==2.9.10
python-dotenv==1.0.1
Requests==2.32.3
